Return the results of the `.src()`
Gulp tasks do their dependency graph blocking on the return value from `gulp.src()` or whatever.  If those tasks aren't returned, things depending on them will proceed immediately.  This can be a problem for report gathering steps, and for steps which should not proceed until the tests are known to pass (such as auto-publishing.)

Because tasks are often copy and pasted blindly from `readme`s, it seems like these tasks should be ready for that sort of thing.